YouTube Landing Page (#39)
This will create the intro and last slide for the website. The user chooses a video, which is then passed to the Home component to be used by the rest of the content in the carousel. After the video is chosen, the carousel is shown. After the carousel is finished, it switches to the final slide of the intro.